Kept only ci_sanity tests for analytics tests
